Update packaging-air-gap-excluding-minio.md
Changing some of the verbiage around who to contact; I think it is OK if we are a bit ambiguous (IE not name the specific replicated titles). The hope is that they reach out to replicated in general - if it comes in from their AE or CSM or CE or TAM it shouldn't be too impactful. having said that, if we want to be specific that is OK too but technically the CE is 'customer engineer' I think (using this for reference https://wiki.replicated.io/doc/customer-engineering-Gwwq2L91i2 )