HOLD ON RELEASE - 46021: Exclude MinIO from air gap bundles (#312)
* Add topic to exclude MinIO from air gap bundles

* Edit

* Edits

* Edits

* minor wording edit

* move out of packaging

* Update packaging-air-gap-excluding-minio.md

Changing some of the verbiage around who to contact; I think it is OK if we are a bit ambiguous (IE not name the specific replicated titles). The hope is that they reach out to replicated in general - if it comes in from their AE or CSM or CE or TAM it shouldn't be too impactful. having said that, if we want to be specific that is OK too but technically the CE is 'customer engineer' I think (using this for reference https://wiki.replicated.io/doc/customer-engineering-Gwwq2L91i2 )

* Edits

Co-authored-by: Paige Calvert <paige@replicated.com>
Co-authored-by: Justin Nordeste <84782406+JustinN-Replicated@users.noreply.github.com>