doc: Clarifies the recommendations to not lead users to make bad design choices by looking the docs examples
## Description 

The example to use `kots.io/exclude` might lead users to think that is recommended they try to install what is installed by KURL when it should not be done and would be considered a bad practice. 

So, the goal of this PR is added the info to let the users be able to carefully think about how they will use this feature.