Update replicated-onboarding.md
Put headers above the code blocks.

Rationale: when I'm skimming a doc looking for the code to run, I am probably not reading 3-4 lines of text explaining what the line of code does, and when I would or would not want to run it. Since this is a fork in the logic, I would suggest we indicate with headers what the code block in each section is going to do, so it can't be missed.