Update kURL to clarify version expectation
The current link to "creating a kURL installer" doesn't actually help me find the info I need. I need the kURL installer ID slug. Outside of trying to use kURL.sh to get this ID, I need to go find it in my installers history. Still not the easiest to discover ,but a small improvement towards pointing me in the right direction. 

I'm not sure if this page is deep linkable or not since it's context specific, but this is an example of what the URL for the right page looks like https://vendor.replicated.com/apps/{app-id}/channels/{channel-id)/kubernetes-installers?page=0