Documents new custom field types (#3545)
TL;DR
-----

Adds the new types of custom license fields

Details
-------

Updates the docs on custom license fields to include the "Password"
and "Enum" types. Passwords are eseentially strings that are
considered sufficiently sensitive to be masked in the UI, while enums
allow you to specify a list of options for the users to choose from.

Provides more description of each type by moving the list of types
into its own section and putting the types in a table with the name,
an explanation, and the type of widget used on the "Manage Customer"
screen.

Co-authored-by: Paige Calvert <paige@replicated.com>